requirements.txt
aiofiles==24.1.0
aiogram==3.16.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
APScheduler==3.6.3
asgiref==3.7.2
attrs==24.3.0
beautifulsoup4==4.12.3
cachetools==4.2.2
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
decorator==5.1.1

frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
instaloader==4.14.1
magic-filter==1.0.12
multidict==6.1.0
my-app==0.1.0
prayer-times-calculator==0.0.12
propcache==0.2.1
py==1.11.0
pydantic==2.10.4
pydantic_core==2.27.2
python-telegram-bot==13.15
pytube==15.0.0
pytz==2024.2
requests==2.32.3
retry==0.9.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.4.4
style==1.1.0
tornado==6.1
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
update==0.0.1
urllib3==2.2.3
yarl==1.18.3
